requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
click==8.1.7
python-dotenv==1.0.1
